Add cursory locking of the if_snd queue to prevent races between the ifq empty check and processing the first packet (or, worse, double-processing.)
There needs to be more locking of the upper-level structures in this driver, but
this is a good place to start.  All of the things below the ifnet have their
own locking or don't need it.  (Especially where there are already critical
sections to prevent migration when accessing per-CPU registers.)

Reported by:	stas