Begin separating out the "setup descriptor", "setup descriptor list" and "handoff" from the actual TX path.
Packets going onto a software TX queue will need to be setup as per
normal but won't be added to the hardware TX queue until later.

Details:

* ath_tx_start() is now ath_tx_normal_setup() - it sets up the packet,
  configures the first descriptor but doesn't add it to the TX queue
  via ath_tx_handoff();
* ath_tx_handoff() now doesn't call ath_tx_chaindesclist() to chain
  together descriptors as required by the MAC;

* a new ath_tx_start() now:

  + determines the destination TX queue;
  + calls ath_tx_normal_setup() to do the packet and descriptor setup;
  + calls ath_tx_chaindesclist() to chain the descriptors together
    for the MAC;
  + calls ath_tx_handoff() to dispatch to the hardware.

The packets going onto the mcast software TXQ (which is implemented
just like a hardware TXQ, but with no hardware dispatch :) already do
this but the decision not to dispatch them directly is done in
ath_tx_handoff() which is something that needs to be uncoupled a bit better.

Whilst I'm here (and having to do development on an 80x50 VGA console,
due to needing to interpret kernel panics) undo some of my brain damaged
non-style(9) commits.