Squirrel away a cheap copy of the sequence number to use in later trechery.
When doing A-MPDU, the sequence number of the current packet is needed
in order to ensure what's queued to the hardware is within the BAW.
When doing aggregation, the sequence number of the current packet is
needed to see whether to squeeze said packet into an aggregate.

The code could use M_SEQNO() (and I may do that later) but one of my aims
is to separate out the code which handles mbufs to the code which does
all the rest of it.