This deviates a bit from the reference code; I may end up backing this out at a later stage.
This unifies the "is this packet supposed to be considered as part of the
BAW?" logic. I've added a bit in the ath_buf.bf_state struct which indicates
whether it's going to be considered for inclusion or not.

Packets that aren't part of the BAW calculation are thus not passed to the
BAW add or update code.

Eventually, the aggregate forming code will also not include this packet
as part of an aggregate and will be sent as a non-aggregate packet.

This fixes the hangs, interface resets and general unhappiness that has been
happening with my recent changes.