* Disable BAR sending for now, I'm still not entirely convinced that   it's all setup correctly and the right BAW left edge value is being   sent (ie, I think I need to add some locking to ensure that new values   aren't allocated by the TX send context, as completion and TX occur   in different thread/processes.)
  I'll revisit correct BAR TX'ing later on.

* Add/change some debugging around to make tracking down strange sequence
  number issues easier.

And the actual bugfix:

* re-schedule the TID for more TX scheduling after the aggregate
  completion handler is run - either the error handler (which retries
  all subframes if possible) or after the completion handler itself.
  In both instances, since packets may be requeued on the software
  queue, the TID needs rescheduling.

Without this, I noticed the throughput would drop immediately when
any retries occured, and lots of frames were being scheduled later
on when I was issuing a single ICMP ping. That ping was causing the
software queue to be scheduled, and lots of frames were thus being
re-scheduled to the hardware (and updating the BAW in the process.)

With this, things are stable. Slow (same as non-aggregate speed)
and the software retry rate is too high for my liking, but packets
now seem to zip along.