Begin refactoring out the rts/cts rate/duration code, in preparation for supporting delayed rate lookups.
The intention is to support doing a rate lookup just before the packet
is queued to the hardware (if required), so software retransmits do not
simply use the old rate.

This isn't yet very well tested, either on legacy or current NICs.
Some further work is needed to enable 11n protection, as well as signaling
whether a rate lookup is required or not (as there may be static information
passed in to ath_tx_raw_xmit().)