Add in a (temporary!) hack to allow the CCA to be forced.
I'm seeing some very bizarre merlin behaviour where it thinks the
air is > 95% busy, and thus never thinks the air is ready for TX.
Comparing the RX clear / RX frame counters to Sowl (AR9160) show that
they're decoding the same number of frames, but Merlin seems to
calibrate the NF much, much lower than Sowl - and then thinks it
can't TX for most of it.

When I force the CCA to be the same as what Sowl sees (ie, ~ -85dB)
then suddenly it thinks it can TX.

I'll chase up _why_ this happens with the Atheros baseband/radio
team.