Introduce a new variable, TCPATH which overrides other path elements during a full world build.  This allows portions of the toolchain to be replaced with external components.  It has been tested with cc, c++, nm, and ld replaced.  TCPATH is used by installed the tools you with to use under their common names and then building world with a command like:
	make TCPATH=/tmp/tc buildworld

TCPATH support depends on the ability to call cc/c++ with --sysroot
which in turn requires that the linker the compiler users (hard
coded at compile time) also support --sysroot.  In the case of GNU
ld this is non-default.  You can produce a linker that works exactly
like normal but also supports --sysroot by building binutils with
the configure argument "--with-sysroot=/".

The current patch is quite inefficient.   It still builds all crossbuild
tools including copies of gcc and clang.  A more efficient and selective
approach is certainly possible.