Several style type changes.
Main functional change is not to run events from the prior boot or
when the driver is shutting down.  Tested added and removing JBOD
drives.  The driver adds and remove drives.  A sysctl
hw.mfi.detect_jbod_change enables or disables this.  it is set by
default to enable this.  On boot only act on events from this
boot.  This avoids stale disk events stating disks have gone
and come back from.  This happens when the kernel doesn't shutdown
the card (ie. power cycle).

Switch hw.mfi.msi to 1 by default.  It works better on the cards
I've tried.

This driver should be very close to being merged into the main
FreeBSD tree now.