Add a test case for each FIB to properly populate on traffic, checking both netstat -rnf inet6 and ndp -an for the neighbor.
Add a) longer delays and b) clear the neighbor table between each test
case to make stable/7 testing a lot more deterministic and less exposed
to network stack address list locking, etc. races.  It's very close to
be deterministic in all cases with that.

Adjust the expected return codes for the few cases that behave differently
in 7 to 8 and later.

Update documentation for the new test and correct some spelling errors.

Sponsored by:	Cisco Systems, Inc.