This is a first cut at publishing heavily WIP work to port FreeBSD 64bit to run as a XEN paravirtualised kernel.
It is very early days and therefore only has a guarantee of
compilability on the branch. The following have been deferred for
later review:

 - style nits.
 - comments

What would be really helpful for review would be:
 - design errors
 - bugs
 - file organisation, API use etc.

The primary focus of the branch at this point is functionality.

We use a config(5) option:
options	 	NATIVE

to indicate "baremetal" support. The intent of this option is to
exclude files from the PV kernel that only the "baremetal" kernel
needs. This is similar to how the i386 port implements this. An
alternative method is needed before merge to -current

Reviewed by: attilio, gibbs

Approved by: gibbs