Switch the syscall interface on EABI.
The syscall we are after is now passed in r7. The value, while still placed
in the instruction, is unused. r7 was chosen as it is what Linux uses and
there is no point being different.

This should help someone add Thumb support as we can now express all
possible syscall values. The Thumb swi instruction only has 8 bits for the
the syscall value.