The following changes go into this diff:
i) Add copyright information from the native file
ii) Rename some variables to be more accurately descriptive of their function.
iii) Add a boottime implementation of pmap_kenter() which uses mmu_map.h

Please note that this is heavily wip and therefore probably very broken.
The idea behind this commit is to publish the changes so that the branch can be
pulled up to HEAD, and progress noted.

The ideal way to get pmap operating would be to use 2MB or greater page sized
pages to direct map all the physical/machine pages and use direct map. This is
slightly more complicated by the fact that xen disallows write access to pages
that are part of the page table hierarchy and therefore the native implementation
of dmap will not work directly.

Approved by:	gibbs (implicit)