- Fix a LOR deadlock dropping the callout lock while executing the handler directly from hardware interrupt context.
- Add migration support for callouts which runs from hw interrupt context.

- Refactor a couple of comments to reflect the new world order.

TODO: implement proper stats for direct callouts.

Just a note: I'd like to thank flo@ for his invaluable help in testing and
issues, mav@ that helped me in tackling the problem and Yahoo! which
provided access to the machines on which tests were run.

Reported by:		avg, flo [1]
Discused with:		mav